separate code template and code generation tests
